requirements.txt
edge-tts
discord.py[voice]
python-dotenv
